move synchronization exceptions from api to impl (#171)
