+++ data/crac-file/crac-impl/src/test/java/com/farao_community/farao/data/crac_impl/threshold/RelativeFlowThresholdTest.java
@@ -9,6 +9,8 @@ package com.farao_community.farao.data.crac_impl.threshold;
 
 import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.data.crac_api.*;
+import com.farao_community.farao.data.crac_impl.AlreadySynchronizedException;
+import com.farao_community.farao.data.crac_impl.NotSynchronizedException;
 import com.powsybl.iidm.import_.Importers;
 import com.powsybl.iidm.network.Network;
 import org.junit.Before;
